10/23/25 Change log:
Phewwww this took me sooooo long to do. Who would have thought removing an item from a dictionary would be so time-consuming? To start off with the updates and changes:

- *Added deal_Card() function. Cards are now pulled using pull_Card(), then printed to the console using deal_Card() (To tell the player what card they pulled!) This function also handles the removing of the card from the deck using the .pop() function!*

- *Added a couple debugging functions (deck_Count(), clear_Deck())*

- *Removed unnecessary imports in main*

- *More comments! More explanations!*

- *Simplified function names*

I can't wait to get back to working on this project again! It's a lot of fun even if I'm up at 4am writing this!

~Swrd